Aggregate 1 - encapsulation added
