+++ COVID19Py/getData.py
@@ -1,45 +1,133 @@
 from typing import Dict, List
-from .covid19 import COVID19
+import requests
+import json
 
 
 class GetData(object):
-
     default_url = "https://covid-tracker-us.herokuapp.com"
+    url = ""
+    data_source = ""
+    previousData = None
+    latestData = None
+    _valid_data_sources = []
 
-    def __init__(self, url=default_url, data_source='jhu'):
-      self.url = url
-      self.data_source = data_source
-      self.obj = COVID19(self.url, self.data_source)
+    mirrors_source = "https://raw.github.com/Kamaropoulos/COVID19Py/master/mirrors.json"
+    mirrors = None
+
+    def __init__(self, url="https://covid-tracker-us.herokuapp.com", data_source='jhu'):
+        # Skip mirror checking if custom url was passed
+        if url == self.default_url:
+            # Load mirrors
+            response = requests.get(self.mirrors_source)
+            response.raise_for_status()
+            self.mirrors = response.json()
+
+            # Try to get sources as a test
+            for mirror in self.mirrors:
+                # Set URL of mirror
+                self.url = mirror["url"]
+                result = None
+                try:
+                    result = self._getSources()
+                except Exception as e:
+                    # URL did not work, reset it and move on
+                    self.url = ""
+                    continue
+
+                # TODO: Should have a better health-check, this is way too hacky...
+                if "jhu" in result:
+                    # We found a mirror that worked just fine, let's stick with it
+                    break
+
+                # None of the mirrors worked. Raise an error to inform the user.
+                raise RuntimeError("No available API mirror was found.")
+
+        else:
+            self.url = url
+
+        self._valid_data_sources = self._getSources()
+        if data_source not in self._valid_data_sources:
+            raise ValueError(
+                "Invalid data source. Expected one of: %s" % self._valid_data_sources)
+        self.data_source = data_source
+
+    def _update(self, timelines):
+        latest = self.getLatest()
+        locations = self.getLocations(timelines)
+        if self.latestData:
+            self.previousData = self.latestData
+        self.latestData = {
+            "latest": latest,
+            "locations": locations
+        }
+
+    def _getSources(self):
+        response = requests.get(self.url + "/v2/sources")
+        response.raise_for_status()
+        return response.json()["sources"]
+
+    def _request(self, endpoint, params=None):
+        if params is None:
+            params = {}
+        response = requests.get(self.url + endpoint,
+                                {**params, "source": self.data_source})
+        response.raise_for_status()
+        return response.json()
 
     def getAll(self, timelines=False):
-        """
-        :param timelines: Whether timeline information should be returned as well.
-        :return: all Covid data numbers from around the world 
-        """
-       self.obj.getAll(timelines=timelines)
-
+        self._update(timelines)
+        return self.latestData
 
     def getLatestChanges(self):
-          """
-        :return: The difference between the current and the previously fetched covid numbers.
-        """
-        return self.obj.getLatestChanges()
+        changes = None
+        if self.previousData:
+            changes = {
+                "confirmed": self.latestData["latest"]["confirmed"] - self.latestData["latest"]["confirmed"],
+                "deaths": self.latestData["latest"]["deaths"] - self.latestData["latest"]["deaths"],
+                "recovered": self.latestData["latest"]["recovered"] - self.latestData["latest"]["recovered"],
+            }
+        else:
+            changes = {
+                "confirmed": 0,
+                "deaths": 0,
+                "recovered": 0,
+            }
+        return changes
 
     def getLatest(self) -> List[Dict[str, int]]:
-          """
-        :return: The latest total confirmed,deaths and recovered case numbers.
         """
-        return self.obj.getLatest()
+        :return: The latest amount of total confirmed cases, deaths, and recoveries.
+        """
+        data = self._request("/v2/latest")
+        return data["latest"]
 
     def getLocations(self, timelines=False, rank_by: str = None) -> List[Dict]:
-         """
+        """
         Gets all locations affected by COVID-19, as well as latest case data.
         :param timelines: Whether timeline information should be returned as well.
         :param rank_by: Category to rank results by. ex: confirmed
         :return: List of dictionaries representing all affected locations.
         """
-        return self.obj.getLocations(timelines=timelines, rank_by=rank_by)
-       
+        data = None
+        if timelines:
+            data = self._request(
+                "/v2/locations", {"timelines": str(timelines).lower()})
+        else:
+            data = self._request("/v2/locations")
+
+        data = data["locations"]
+
+        ranking_criteria = ['confirmed', 'deaths', 'recovered']
+        if rank_by is not None:
+            if rank_by not in ranking_criteria:
+                raise ValueError(
+                    "Invalid ranking criteria. Expected one of: %s" % ranking_criteria)
+
+            ranked = sorted(
+                data, key=lambda i: i['latest'][rank_by], reverse=True)
+            data = ranked
+
+        return data
 
     def getLocationByCountryCode(self, country_code, timelines=False) -> List[Dict]:
         """
@@ -47,8 +135,14 @@ class GetData(object):
         :param timelines: Whether timeline information should be returned as well.
         :return: A list of areas that correspond to the country_code. If the country_code is invalid, it returns an empty list.
         """
-        country_code = country_code
-        return self.obj.getLocationByCountryCode(country_code=country_code, timelines=timelines)
+        data = None
+        if timelines:
+            data = self._request(
+                "/v2/locations", {"country_code": country_code, "timelines": str(timelines).lower()})
+        else:
+            data = self._request(
+                "/v2/locations", {"country_code": country_code})
+        return data["locations"]
 
     def getLocationByCountry(self, country, timelines=False) -> List[Dict]:
         """
@@ -56,13 +150,18 @@ class GetData(object):
         :param timelines: Whether timeline information should be returned as well.
         :return: A list of areas that correspond to the country name. If the country is invalid, it returns an empty list.
         """
-    
-        return self.obj.getLocationByCountry(country_code=country_code, timelines=timelines)
+        data = None
+        if timelines:
+            data = self._request(
+                "/v2/locations", {"country": country, "timelines": str(timelines).lower()})
+        else:
+            data = self._request("/v2/locations", {"country": country})
+        return data["locations"]
 
     def getLocationById(self, country_id: int):
         """
         :param country_id: Country Id, an int
         :return: A dictionary with case information for the specified location.
         """
-   
-        return self.obj.getLocationById(country_id=country_id)
+        data = self._request("/v2/locations/" + str(country_id))
+        return data["location"]
